minor refactor to move BaseHandlerTest to root package
