+++ src/e2e/java/uk/gov/hmcts/reform/sscs/functional/handlers/fenoaction/FeNoActionHandlerTest.java
@@ -11,7 +11,7 @@ import org.springframework.test.context.junit4.SpringRunner;
 import uk.gov.hmcts.reform.sscs.ccd.domain.EventType;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
 import uk.gov.hmcts.reform.sscs.ccd.domain.State;
-import uk.gov.hmcts.reform.sscs.functional.handlers.uploaddocument.BaseHandlerTest;
+import uk.gov.hmcts.reform.sscs.functional.handlers.BaseHandlerTest;
 
 @RunWith(SpringRunner.class)
 @TestPropertySource(locations = "classpath:config/application_e2e.properties")
